Move POPCs from method to destructor
These were moved from the destructor to a method because using a
destructor brings in the exception handler code at link time resulting
in the code bloat issue (fails to load on versim).

The POPCs actually must occur in a destructor so that jumps (continue,
break, etc) pop correctly.

 Now that -fno-except is used and execption handling is gutted in the compiler,
move these back.