Remove some unneeded constructors from class vDReg
Note that the conditional operators are also duplicated and likely
should be removed, but the specialized versions (eg, taking a float
and s2vFloat16) load a register after the conditional while the
generic versions (taking a vFloat) load the register before the
conditional which results in out of registers in some test instances.
Leaving these for now.