Bug fix: liveness assignments
The wrapper wasn't emiting "assign_lv" for assignments, particularly
in the +=, *=, etc operators resulting in liveness violations in some
cases.  This is a little tricky since, for example, &&= and ||= don't
need live assignments since they take the dst as src and don't have
live versions of their instructions (using a live assign in this case
results in extraneous SFPMOVs).