Rework LReg assigner
The old implementation used an array of pointers to vectors which
caused problems in some instances w/ large fns where gcc's sra pass
failed to remove all of the memory references leading to register
spilling issues.

New version has a new API which makes the user responsible for
preserving the value at the end of the fn.  Arguably this is better as
it lets a register be assigned and not preserved and the interface is
cleaner.  I'm bummed to lose the automatic-ness of the preservation,
though.  Hmm.